Reworked InventorySet, navigation and views. Change view names to "normal"
