+++ .che/project.json
@@ -1,1 +1,1 @@
-{"type":"sap.web","builders":{"configs":{}},"runners":{"configs":{}},"attributes":{"sap.watt.common.setting":["{\"dataBinding\":{\"/webapp/view/security.view.xml\":{\"entitySet\":\"\"}},\"projectType\":[\"sap.watt.uitools.ide.web\",\"sap.watt.uitools.ide.fiori\"],\"mockpreview\":{\"mockUri\":\"http://localhost:8081/mobinv/services/odata.svc/\",\"metadataFilePath\":\"localService/metadata.xml\",\"loadJSONFiles\":false,\"loadCustomRequests\":false,\"mockRequestsFilePath\":\"\"},\"hcpdeploy\":{\"account\":\"p491054trial\",\"name\":\"smartinventorymik\",\"lastVersionWeTriedToDeploy\":\"1.0.0\"}}"]},"mixinTypes":[]}
+{"type":"sap.web","builders":{"configs":{}},"runners":{"configs":{}},"attributes":{"sap.watt.common.setting":["{\"dataBinding\":{\"/webapp/view/security.view.xml\":{\"entitySet\":\"\"}},\"projectType\":[\"sap.watt.uitools.ide.web\",\"sap.watt.uitools.ide.fiori\"],\"mockpreview\":{\"mockUri\":\"http://localhost:8081/mobinv/services/odata.svc/\",\"metadataFilePath\":\"localService/metadata.xml\",\"loadJSONFiles\":false,\"loadCustomRequests\":false,\"mockRequestsFilePath\":\"\"},\"hcpdeploy\":{\"account\":\"p491054trial\",\"name\":\"smartinventorymik\",\"lastVersionWeTriedToDeploy\":\"1.0.0\"},\"layout\":{\"snapshot\":{\"viewport\":{\"x\":-838,\"y\":-384,\"scale\":1},\"collapsedPorts\":[],\"expandedShapes\":[],\"selection\":[\"_minvui5_webapp_view_Inventory_view_xml\"],\"focus\":\"_minvui5_webapp_view_Inventory_view_xml\"}}}"]},"mixinTypes":[]}
