--- webapp/controller/Inventory_S1.controller.js
@@ -1,304 +1,0 @@
-sap.ui.define(["sap/ui/core/mvc/Controller",
-	"sap/m/MessageBox",
-	"./utilities",
-	"sap/ui/core/routing/History",
-	"sap/m/MessageToast"
-], function(BaseController, MessageBox, Utilities, History, MessageToast) {
-	"use strict";
-
-	return BaseController.extend("com.sap.build.standard.smartStore.controller.Inventory_S1", {
-		handleRouteMatched: function(oEvent) {
-			var sAppId = "App5cc13d57ba792f153b22e1fb";
-
-			var oParams = {};
-
-			if (oEvent.mParameters.data.context) {
-				this.sContext = oEvent.mParameters.data.context;
-
-			} else {
-				if (this.getOwnerComponent().getComponentData()) {
-					var patternConvert = function(oParam) {
-						if (Object.keys(oParam).length !== 0) {
-							for (var prop in oParam) {
-								if (prop !== "sourcePrototype") {
-									return prop + "(" + oParam[prop][0] + ")";
-								}
-							}
-						}
-					};
-
-					this.sContext = patternConvert(this.getOwnerComponent().getComponentData().startupParameters);
-
-				}
-			}
-
-			var oPath;
-
-			if (this.sContext) {
-				oPath = {
-					path: "/" + this.sContext,
-					parameters: oParams
-				};
-				this.getView().bindObject(oPath);
-			}
-
-		},
-		_onPageNavButtonPress: function(oEvent) {
-
-			var oBindingContext = oEvent.getSource().getBindingContext();
-
-			return new Promise(function(fnResolve) {
-
-				this.doNavigate("7f3446de6ee0a8ea0d67ae610_S0", oBindingContext, fnResolve, "");
-			}.bind(this)).catch(function(err) {
-				if (err !== undefined) {
-					MessageBox.error(err.message);
-				}
-			});
-
-		},
-		doNavigate: function(sRouteName, oBindingContext, fnPromiseResolve, sViaRelation) {
-			var sPath = (oBindingContext) ? oBindingContext.getPath() : null;
-			var oModel = (oBindingContext) ? oBindingContext.getModel() : null;
-
-			var sEntityNameSet;
-			if (sPath !== null && sPath !== "") {
-				if (sPath.substring(0, 1) === "/") {
-					sPath = sPath.substring(1);
-				}
-				sEntityNameSet = sPath.split("(")[0];
-			}
-			var sNavigationPropertyName;
-			var sMasterContext = this.sMasterContext ? this.sMasterContext : sPath;
-
-			if (sEntityNameSet !== null) {
-				sNavigationPropertyName = sViaRelation || this.getOwnerComponent().getNavigationPropertyForNavigationWithContext(sEntityNameSet, sRouteName);
-			}
-			if (sNavigationPropertyName !== null && sNavigationPropertyName !== undefined) {
-				if (sNavigationPropertyName === "") {
-					this.oRouter.navTo(sRouteName, {
-						context: sPath,
-						masterContext: sMasterContext
-					}, false);
-				} else {
-					oModel.createBindingContext(sNavigationPropertyName, oBindingContext, null, function(bindingContext) {
-						if (bindingContext) {
-							sPath = bindingContext.getPath();
-							if (sPath.substring(0, 1) === "/") {
-								sPath = sPath.substring(1);
-							}
-						} else {
-							sPath = "undefined";
-						}
-
-						// If the navigation is a 1-n, sPath would be "undefined" as this is not supported in Build
-						if (sPath === "undefined") {
-							this.oRouter.navTo(sRouteName);
-						} else {
-							this.oRouter.navTo(sRouteName, {
-								context: sPath,
-								masterContext: sMasterContext
-							}, false);
-						}
-					}.bind(this));
-				}
-			} else {
-				this.oRouter.navTo(sRouteName);
-			}
-
-			if (typeof fnPromiseResolve === "function") {
-				fnPromiseResolve();
-			}
-
-		},
-		_onTableItemPress: function(oEvent) {
-
-			var oBindingContext = oEvent.getParameter("listItem").getBindingContext();
-
-			return new Promise(function(fnResolve) {
-				this.doNavigate("AlertInfo_S2", oBindingContext, fnResolve, "");
-			}.bind(this)).catch(function(err) {
-				if (err !== undefined) {
-					MessageBox.error(err.message);
-				}
-			});
-
-		},
-		_onTableItemPress1: function(oEvent) {
-
-			var oBindingContext = oEvent.getParameter("listItem").getBindingContext();
-
-			return new Promise(function(fnResolve) {
-				this.doNavigate("AlertInfo_S2", oBindingContext, fnResolve, "");
-			}.bind(this)).catch(function(err) {
-				if (err !== undefined) {
-					MessageBox.error(err.message);
-				}
-			});
-
-		},
-		_onTableItemPress2: function(oEvent) {
-
-			var oBindingContext = oEvent.getParameter("listItem").getBindingContext();
-
-			return new Promise(function(fnResolve) {
-				this.doNavigate("AlertInfo_S2", oBindingContext, fnResolve, "");
-			}.bind(this)).catch(function(err) {
-				if (err !== undefined) {
-					MessageBox.error(err.message);
-				}
-			});
-
-		},
-		_onTableItemPress3: function(oEvent) {
-
-			var oBindingContext = oEvent.getParameter("listItem").getBindingContext();
-
-			return new Promise(function(fnResolve) {
-				this.doNavigate("AlertInfo_S2", oBindingContext, fnResolve, "");
-			}.bind(this)).catch(function(err) {
-				if (err !== undefined) {
-					MessageBox.error(err.message);
-				}
-			});
-
-		},
-		
-		_onStockChange: function(oEvent) {
-			console.log("Stock value changed to '" + oEvent.getParameter("value") + "'");
-		},
-		
-		updateBindingOptions: function(sCollectionId, oBindingData, sSourceId) {
-			this.mBindingOptions = this.mBindingOptions || {};
-			this.mBindingOptions[sCollectionId] = this.mBindingOptions[sCollectionId] || {};
-
-			var aSorters = this.mBindingOptions[sCollectionId].sorters;
-			var aGroupby = this.mBindingOptions[sCollectionId].groupby;
-
-			// If there is no oBindingData parameter, we just need the processed filters and sorters from this function
-			if (oBindingData) {
-				if (oBindingData.sorters) {
-					aSorters = oBindingData.sorters;
-				}
-				if (oBindingData.groupby || oBindingData.groupby === null) {
-					aGroupby = oBindingData.groupby;
-				}
-				// 1) Update the filters map for the given collection and source
-				this.mBindingOptions[sCollectionId].sorters = aSorters;
-				this.mBindingOptions[sCollectionId].groupby = aGroupby;
-				this.mBindingOptions[sCollectionId].filters = this.mBindingOptions[sCollectionId].filters || {};
-				this.mBindingOptions[sCollectionId].filters[sSourceId] = oBindingData.filters || [];
-			}
-
-			// 2) Reapply all the filters and sorters
-			var aFilters = [];
-			for (var key in this.mBindingOptions[sCollectionId].filters) {
-				aFilters = aFilters.concat(this.mBindingOptions[sCollectionId].filters[key]);
-			}
-
-			// Add the groupby first in the sorters array
-			if (aGroupby) {
-				aSorters = aSorters ? aGroupby.concat(aSorters) : aGroupby;
-			}
-
-			var aFinalFilters = aFilters.length > 0 ? [new sap.ui.model.Filter(aFilters, true)] : undefined;
-			return {
-				filters: aFinalFilters,
-				sorters: aSorters
-			};
-
-		},
-		createFiltersAndSorters: function() {
-			this.mBindingOptions = {};
-			var oBindingData, aPropertyFilters;
-			oBindingData = {};
-			oBindingData.filters = [];
-			aPropertyFilters = [];
-
-			aPropertyFilters.push(new sap.ui.model.Filter("ItemType", "EQ", "Non-Perishable"));
-			oBindingData.filters.push(new sap.ui.model.Filter(aPropertyFilters, false));
-
-			this.updateBindingOptions("sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-3-content-build_simple_Table-1489097986245", oBindingData);
-			oBindingData = {};
-			oBindingData.filters = [];
-			aPropertyFilters = [];
-
-			aPropertyFilters.push(new sap.ui.model.Filter("ItemType", "EQ", "Perishable"));
-			oBindingData.filters.push(new sap.ui.model.Filter(aPropertyFilters, false));
-
-			this.updateBindingOptions("sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-5-content-build_simple_Table-1489098016833", oBindingData);
-			oBindingData = {};
-			oBindingData.filters = [];
-			aPropertyFilters = [];
-
-			aPropertyFilters.push(new sap.ui.model.Filter("Status", "EQ", "Re-Stock"));
-			aPropertyFilters.push(new sap.ui.model.Filter("Status", "EQ", "Deteriorating"));
-			oBindingData.filters.push(new sap.ui.model.Filter(aPropertyFilters, false));
-
-			this.updateBindingOptions("sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-1487204305428-content-build_simple_Table-1489098372093", oBindingData);
-
-		},
-		applyFiltersAndSorters: function(sControlId, sAggregationName, chartBindingInfo) {
-			if (chartBindingInfo) {
-				var oBindingInfo = chartBindingInfo;
-			} else {
-				var oBindingInfo = this.getView().byId(sControlId).getBindingInfo(sAggregationName);
-			}
-			var oBindingOptions = this.updateBindingOptions(sControlId);
-			this.getView().byId(sControlId).bindAggregation(sAggregationName, {
-				model: oBindingInfo.model,
-				path: oBindingInfo.path,
-				parameters: oBindingInfo.parameters,
-				template: oBindingInfo.template,
-				templateShareable: true,
-				sorter: oBindingOptions.sorters,
-				filters: oBindingOptions.filters
-			});
-
-		},
-		onInit: function() {
-			this.oRouter = sap.ui.core.UIComponent.getRouterFor(this);
-			this.oRouter.getTarget("Inventory_S1").attachDisplay(jQuery.proxy(this.handleRouteMatched, this));
-			this.mAggregationBindingOptions = {};
-			this.createFiltersAndSorters();
-
-			this.applyFiltersAndSorters("sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-3-content-build_simple_Table-1489097986245", "items");
-
-			this.applyFiltersAndSorters("sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-5-content-build_simple_Table-1489098016833", "items");
-
-			this.applyFiltersAndSorters("sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-1487204305428-content-build_simple_Table-1489098372093", "items");
-
-		},
-		onExit: function() {
-
-			// to destroy templates for bound aggregations when templateShareable is true on exit to prevent duplicateId issue
-			var aControls = [{
-				"controlId": "sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-1-content-build_simple_Table-1",
-				"groups": ["items"]
-			}, {
-				"controlId": "sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-3-content-build_simple_Table-1489097986245",
-				"groups": ["items"]
-			}, {
-				"controlId": "sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-5-content-build_simple_Table-1489098016833",
-				"groups": ["items"]
-			}, {
-				"controlId": "sap_Responsive_Page_0-content-sap_m_IconTabBar-1487204151231-items-sap_m_IconTabFilter-1487204305428-content-build_simple_Table-1489098372093",
-				"groups": ["items"]
-			}];
-			for (var i = 0; i < aControls.length; i++) {
-				var oControl = this.getView().byId(aControls[i].controlId);
-				if (oControl) {
-					for (var j = 0; j < aControls[i].groups.length; j++) {
-						var sAggregationName = aControls[i].groups[j];
-						var oBindingInfo = oControl.getBindingInfo(sAggregationName);
-						if (oBindingInfo) {
-							var oTemplate = oBindingInfo.template;
-							oTemplate.destroy();
-						}
-					}
-				}
-			}
-
-		}
-	});
-}, /* bExport= */ true);
