+++ webapp/i18n/i18n_en.properties
@@ -1,6 +1,6 @@
 
 # XTIT:Application name
-appTitle=Smart Store
+appTitle=Mobile Inventory
 
 # YDES:Application description
 appDescription=
@@ -11,14 +11,20 @@ errorTitle=Error
 # YMSG:Error dialog description
 errorText=Request failed
 
+#Inventories header
+invHeaderTitle=Inventory
+
 # Inventories View - table header
 ProductDescription=Product description
 PriceDescription=Price
 CurrencyDescription=Currency
-ShelfStockDescription=Shelf stock
+ShelfStockDescription=Shelf Stock
 LocationDescription=Location
 ImageDescription=Image
 StatusDescription=Status
 
 # Inventories View - table items tooltips
 imageTooltip=Product picture
+
+# ProductInfo header
+productHeaderTitle=Product Informations
