+++ webapp/manifest.json
@@ -54,7 +54,7 @@
 	"sap.ui5": {
 		"_version": "1.3.0",
 		"rootView": {
-			"viewName": "generated.app.view.10fba274c5fcd1720d8123c83_S3",
+			"viewName": "com.sap.build.standard.smartStore.Inventory.view",
 			"type": "XML"
 		},
 		"dependencies": {
@@ -101,82 +101,40 @@
 				"clearTarget": false,
 				"controlAggregation": "pages",
 				"bypassed": {
-					"target": ["Inventory_S1"]
+					"target": []
 				}
 			},
 			"targets": {
-				"AlertInfo_S2": {
+				"Inventory": {
 					"controlAggregation": "pages",
-					"viewName": "AlertInfo_S2",
-					"viewId": "AlertInfo_S2",
-					"viewLevel": 6,
-					"transition": "slide"
-				},
-				"Inventory_S1": {
-					"controlAggregation": "pages",
-					"viewName": "Inventory_S1",
-					"viewId": "Inventory_S1",
-					"viewLevel": 5,
-					"transition": "slide"
-				},
-				"a8138e965e4e48e60d84a5545_S5": {
-					"controlAggregation": "pages",
-					"viewName": "a8138e965e4e48e60d84a5545_S5",
-					"viewId": "a8138e965e4e48e60d84a5545_S5",
-					"viewLevel": 2,
-					"transition": "slide"
-				},
-				"7f3446de6ee0a8ea0d67ae610_S0": {
-					"controlAggregation": "pages",
-					"viewName": "7f3446de6ee0a8ea0d67ae610_S0",
-					"viewId": "7f3446de6ee0a8ea0d67ae610_S0",
+					"viewName": "Inventory",
+					"viewId": "Inventory",
 					"viewLevel": 1,
 					"transition": "slide"
 				},
-				"security": {
-					"controlAggregation": "pages",
-					"viewName": "security",
-					"viewId": "security",
-					"viewLevel": 3,
-					"transition": "slide"
-				},
-				"10fba274c5fcd1720d8123c83_S3": {
-					"controlAggregation": "pages",
-					"viewName": "10fba274c5fcd1720d8123c83_S3",
-					"viewId": "10fba274c5fcd1720d8123c83_S3",
-					"viewLevel": 4,
+				"ProductInfo": {
+					"viewName": "ProductInfo",
+					"viewId": "ProductInfo",
+					"viewLevel": 1,
 					"transition": "slide"
 				}
 			},
-			"routes": [{
-				"pattern": "AlertInfo_S2/:context:",
-				"name": "AlertInfo_S2",
-				"target": ["AlertInfo_S2"]
-			}, {
-				"pattern": "",
-				"name": "Inventory_S1",
-				"target": ["AlertInfo_S2"]
-			}, {
-				"pattern": "Inventory_S1/:context:",
-				"name": "Inventory_S1",
-				"target": ["Inventory_S1"]
-			}, {
-				"pattern": "a8138e965e4e48e60d84a5545_S5/:context:",
-				"name": "a8138e965e4e48e60d84a5545_S5",
-				"target": ["a8138e965e4e48e60d84a5545_S5"]
-			}, {
-				"pattern": "7f3446de6ee0a8ea0d67ae610_S0/:context:",
-				"name": "7f3446de6ee0a8ea0d67ae610_S0",
-				"target": ["7f3446de6ee0a8ea0d67ae610_S0"]
-			}, {
-				"pattern": "security/:context:",
-				"name": "security",
-				"target": ["security"]
-			}, {
-				"pattern": "10fba274c5fcd1720d8123c83_S3/:context:",
-				"name": "10fba274c5fcd1720d8123c83_S3",
-				"target": ["10fba274c5fcd1720d8123c83_S3"]
-			}]
+			"routes": [
+				{
+					"pattern": "",
+					"name": "Inventory",
+					"target": [
+						"Inventory"
+					]
+				},
+				{
+					"name": "ProductInfo",
+					"pattern": "ProductInfo",
+					"target": [
+						"ProductInfo"
+					]
+				}
+			]
 		}
 	},
 	"sap.platform.hcp": {
